Overhaul... again... eliminate the whole pidfile thing and switch to a reconnect-to-managment socket scheme.
The whole idea of keeping pidfiles isn't going to work properly because
the pidfiles don't keep enough information around to be used by subsequent
invokations to reconnect to the existing openvpn management interfaces.

Instead, we'll drop files in /var/tmp/gopenvpn.<conn>.mgmt containing
the TCP port number of the corresponding openvpn's management interface.
Upon (re)start, try to connect to the management ports and consider the
connection up if we can.

NOTE: This is an interim commit and this code DOES NOT WORK yet.