State overhaul.  Enable auto-reconnect.
Rename the RESTART state to RECONNECTING to better indicate its purpose.  The
RECONNECTING state indicates that gopenvpn is trying to re-establish a connection
with an existing openvpn's management interface.

Renumber the states so that they're in the "right" order (i.e. order in which they
normally occur).

The CONNECTING, RECONNECTING and SENTSTATES now mean more-or-less the same thing
insofar as the program's icon is concerned (namely, that it's in some stage of
waiting for a connection).

Remove vpn_applet_autoconnect function and integrate auto-connect with the
the vpn_applet_reconnect_to_mgmt function.