Loop to consume all available input in vpn_config_io_callback.
Previously, vpn_config_io_callback would only read a single line
per call.  On some systems (Fedora 16) and, particularly in batch
mode, it seems there would be multiple lines of input available for
vpn_config_io_callback and if it only consumed a single line, it was never
called again for the unconsumed input.  This way, it's all consumed in
a single call.