Add per-connection timeouts.
A connection that's not established within 60 seconds will be aborted.
The vpn_connect_monitor function is called every second for connections
that are waiting to be established and will abort the connection if it
takes too long.

Harmonize all the config-walking loops.

Move the global "batchmode" flag into the per-applet data structure.

Remove a few more batch mode gtk-isms (I think).